+++ Sources/Location.php
@@ -3,13 +3,13 @@
 namespace Afiqiqmal\SolatJakim\Sources;
 
 use Afiqiqmal\SolatJakim\Library\Constant;
-use Afiqiqmal\SolatJakim\Sources\Model\LocationData;
+use Afiqiqmal\SolatJakim\Sources\Model\ZoneData;
 
 class Location
 {
     /**
      * @param $code
-     * @return LocationData|null
+     * @return ZoneData|null
      */
     public static function getLocationByCode($code)
     {
@@ -46,7 +46,7 @@ class Location
     }
 
     /**
-     * @return LocationData[]
+     * @return ZoneData[]
      */
     public static function getRawData()
     {
@@ -54,10 +54,10 @@ class Location
         $list_extra = array_map('str_getcsv', file(Constant::FILE_LOCATION_EXTRA));
         $data = [];
         foreach ($list as $location) {
-            $data[] = (new LocationData($location[0], $location[1], $location[2], $location[3]));
+            $data[] = (new ZoneData($location[0], $location[1], $location[2], $location[3]));
             foreach ($list_extra as $extra) {
                 if ($extra[1] == $location[3]) {
-                    $data[] = (new LocationData($location[0], $extra[0], $location[2], $location[3]));
+                    $data[] = (new ZoneData($location[0], $extra[0], $location[2], $location[3]));
                 }
             }
         }